# backend/app/Http/Requests/TaskStatusRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     schema="TaskStatusRequest",
 *     type="object",
 *     required={"name", "order"},
 *     @OA\Property(property="name", type="string", maxLength=255),
 *     @OA\Property(property="order", type="integer")
 * )
 */
class TaskStatusRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'name' => 'required|string|max:255',
            'order' => 'required|integer',
        ];
    }
}